BUFIX.py
Fixed an issue where going into the light wouldn't send you to the security office.
Fixed an issue where doing anything in the security office would send the player in an indefinite loop.
Fixed an issue that would crash the program if the player didn't input an integer for the health in the battle trainer. Now it should correctly set the player to the default health (10)
Maybe other changes?